KAN-92 회원탈퇴(회원)-strategy 패턴 리팩터링 완료, 서비스  테스트, 컨트롤러 테스트 완료

+++ user-service/src/main/java/com/buysellgo/userservice/controller/AuthController.java
@@ -88,6 +88,11 @@ public class AuthController {
     @DeleteMapping("/jwt")
     public ResponseEntity<CommonResDto> deleteToken(
             @RequestHeader(value = "Authorization") String accessToken) {
+        // Bearer 접두어 검증
+        if (!accessToken.startsWith(BEARER_PREFIX.getValue())) {
+            throw new IllegalArgumentException("잘못된 Authorization 헤더 형식입니다.");
+        }
+
         String token = accessToken.replace(BEARER_PREFIX.getValue(), "");
         AuthResult<?> result = authService.deleteToken(token);
